fix: Miscellaneous dat requirement and button oddities
Resolves an issue with Fleet Beacon and Arbiter Tribunal requirements. Currently, Arbiter Tribunal only requires Templar Archives, because it's impossbile to check for (Stargate OR Argosy) AND Templar Archives. The same deal goes for the Dropship, which currently requires a Control Tower despite what its tooltip says. If these issues can't be corrected in a timely manner, I'll change the tooltips to more accurately reflect the current limitations.

I was also reminded of an issue with melee gamemodes loading a different set of tech-related data (like which spells were researched by default, for instance) compared to UMS. For the time being, you'll want to run all games in UMS to avoid issues, which limits your options if your goal is to play singleplayer against AI in melee maps. Hopefully it's not too much of a hassle, and hopefully the cause of this bizarre issue will be triaged and resolved shortly.